fix: some reported dependabot issues (#66)
More to come.

Some dependabot issues were caused by eslint. The upgrade resulted in a
number of issues having to be fixed hence the size of this PR.

Added some resolutions. dns-packet will be removed in the next PR which
will upgrade webpack to webpack 5. Hopefully this will eliminate the
remaining dependabot issues.